+++ packages/provider-mssql/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ts-linq/provider-mssql
 
+## 3.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.3.0
+  - @ts-linq/core@3.4.1
+  - @ts-linq/dialect-mssql@2.7.1
+  - @ts-linq/metadata@4.1.2
+
 ## 3.0.18
 
 ### Patch Changes